requirements.txt
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.2.0
pandas==2.2.0
folium==0.15.1
geopy==2.4.1